requirements.txt
Flask
gunicorn
pyyaml
tensorflow-datasets